requirements.txt
pandas
numpy
scikit-learn
django
django-cors-headers
django-autoslug
djangorestframework
python-dotenv
django-crispy-forms
crispy-bootstrap5
